reuse GetUniqueGcsBuckets and update READMEs
