+++ examples/gke-tpu-7x/README.md
@@ -233,6 +233,8 @@ To avoid recurring charges for the resources used, clean up the resources provis
 ./gcluster destroy DEPLOYMENT_NAME
 ```
 
+**Note:** GCS buckets created for Terraform state are not deleted by the `./gcluster destroy` command and must be deleted manually.
+
 ## Appendix
 
 ### Node Count Calculation
@@ -461,5 +463,3 @@ The blueprint includes a sample job (`shared-fs-job`) that demonstrates how two 
     ```
 
 The logs will display content from `shared_output.txt`, showing timestamps and hostnames from both pods, confirming that the filesystem is truly shared.
-
-**Note:** GCS buckets created for Terraform state are not deleted by the `./gcluster destroy` command and must be deleted manually.
