+++ examples/gke-tpu-v6e/README.md
@@ -215,6 +215,8 @@ To avoid recurring charges for the resources used on this page, clean up the res
    ./gcluster destroy gke-tpu-v6e/
    ```
 
+**Note:** GCS buckets created for Terraform state are not deleted by the `./gcluster destroy` command and must be deleted manually.
+
 ## Appendix
 
 ### Useful TPU links
@@ -387,5 +389,3 @@ The blueprint includes a sample job (`shared-fs-job`) that demonstrates how two 
     ```
 
 The logs will display content from `shared_output.txt`, showing timestamps and hostnames from both pods, confirming that the filesystem is truly shared.
-
-**Note:** GCS buckets created for Terraform state are not deleted by the `./gcluster destroy` command and must be deleted manually.
